# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural drawing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drawings:
    get:
      operationId: listDrawings
      tags: [drawings]
      summary: List architectural drawings
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DrawingStatus"
        - name: discipline
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DrawingDiscipline"
      responses:
        "200":
          description: Drawing library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drawing"
    post:
      operationId: createDrawing
      tags: [drawings]
      summary: Add a drawing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingInput"
      responses:
        "201":
          description: Created drawing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
...
  /drawings/{id}:
    get:
      operationId: getDrawing
      tags: [drawings]
      summary: Get a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Drawing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
        "404":
          description: Not found
    patch:
      operationId: updateDrawing
      tags: [drawings]
      summary: Update a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingUpdate"
      responses:
        "200":
          description: Updated drawing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
    delete:
      operationId: deleteDrawing
      tags: [drawings]
      summary: Delete a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "403":
          description: Administrator access required
  /drawings/{id}/assignment:
    patch:
      operationId: updateDrawingAssignment
      tags: [drawings]
      summary: Assign or unassign a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingAssignmentUpdate"
      responses:
        "200":
          description: Updated drawing assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drawing"
        "404":
          description: Drawing not found
  /drawings/{id}/uploads:
    get:
      operationId: listDrawingUploads
      tags: [drawings]
      summary: List upload history for a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Upload history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrawingUpload"
    post:
      operationId: recordDrawingUpload
      tags: [drawings]
      summary: Record a completed drawing file upload
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingUploadInput"
      responses:
        "201":
          description: Recorded upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawingUpload"
        "404":
          description: Drawing not found
  /drawings/{id}/uploads/{uploadId}:
    delete:
      operationId: deleteDrawingUpload
      tags: [drawings]
      summary: Delete a drawing upload and its stored file
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/UploadId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Upload not found
  /drawings/{id}/comments:
    get:
      operationId: listDrawingComments
      tags: [drawings]
      summary: List review comments for a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Review comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrawingComment"
    post:
      operationId: createDrawingComment
      tags: [drawings]
      summary: Leave a review comment on a drawing
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingCommentInput"
      responses:
        "201":
          description: Created review comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawingComment"
        "404":
          description: Drawing not found
  /drawings/{id}/comments/{commentId}:
    patch:
      operationId: updateDrawingComment
      tags: [drawings]
      summary: Edit a drawing review comment
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingCommentUpdate"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawingComment"
        "404":
          description: Comment not found
    delete:
      operationId: deleteDrawingComment
      tags: [drawings]
      summary: Delete a drawing review comment
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CommentId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Comment not found
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
    UploadId:
      name: uploadId
      in: path
      required: true
      schema:
        type: number
        minimum: 1
    CommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: number
        minimum: 1
  schemas:
...
    DrawingStatus:
      type: string
      enum: [draft, in_review, approved, issued, superseded]
    DrawingDiscipline:
      type: string
    Drawing:
      type: object
      properties:
        id:
          type: number
        drawingNumber:
          type: string
        title:
          type: string
        discipline:
          $ref: "#/components/schemas/DrawingDiscipline"
        status:
          $ref: "#/components/schemas/DrawingStatus"
        assignedTo:
          type: ["string", "null"]
        revision:
          type: string
        projectName:
          type: string
        sheetSize:
          type: string
          enum: [A0, A1, A2, A3, A4]
        author:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        issuedDate:
          type: ["string", "null"]
          format: date
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        attachmentPath:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        attachmentSize:
          type: ["number", "null"]
        attachmentContentType:
          type: ["string", "null"]
      required: [id, drawingNumber, title, discipline, status, assignedTo, revision, projectName, sheetSize, author, updatedAt, createdAt]
    DrawingAssignmentUpdate:
      type: object
      properties:
        assigneeName:
          type: ["string", "null"]
        assignedBy:
          type: string
          minLength: 1
      required: [assignedBy]
    DrawingInput:
      type: object
      properties:
        drawingNumber:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        discipline:
          $ref: "#/components/schemas/DrawingDiscipline"
        status:
          $ref: "#/components/schemas/DrawingStatus"
        revision:
          type: string
          minLength: 1
        projectName:
          type: string
          minLength: 1
        sheetSize:
          type: string
          enum: [A0, A1, A2, A3, A4]
        author:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
          format: date
        issuedDate:
          type: string
          format: date
    DrawingUpdate:
      type: object
      properties:
        drawingNumber:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        discipline:
          $ref: "#/components/schemas/DrawingDiscipline"
        status:
          $ref: "#/components/schemas/DrawingStatus"
        revision:
          type: string
          minLength: 1
        projectName:
          type: string
          minLength: 1
        sheetSize:
          type: string
          enum: [A0, A1, A2, A3, A4]
        author:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        issuedDate:
          type: ["string", "null"]
          format: date
        attachmentPath:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        attachmentSize:
          type: ["number", "null"]
        attachmentContentType:
          type: ["string", "null"]
...
    DrawingUpload:
      type: object
      properties:
        id:
          type: number
        drawingId:
          type: number
        filePath:
          type: string
        fileName:
          type: string
        fileSize:
          type: number
        contentType:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, drawingId, filePath, fileName, fileSize, contentType, uploadedBy, uploadedAt]
    DrawingUploadInput:
      type: object
      properties:
        filePath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        fileSize:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
        uploadedBy:
          type: string
          minLength: 1
      required: [filePath, fileName, fileSize, contentType, uploadedBy]
...
    DrawingComment:
      type: object
      properties:
        id:
          type: number
        drawingId:
          type: number
        comment:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, drawingId, comment, author, createdAt]
    DrawingCommentInput:
      type: object
      properties:
        comment:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1
      required: [comment, author]
    DrawingCommentUpdate:
      type: object
      properties:
        comment:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1
...